feat(delete rows): update API with delete endpoint, add delete functionality to frontend
API: create a delete endpoint, refactor code to use routers and clean up autodocs, improve comments
--- FrontEnd: change to primevue table with pagination, add delete buttons and confirmation modal to
each row

BREAKING CHANGE: API endpoints changed slightly, upgraded yarn packages

closes #14

+++ src/backend/mec/schemas.py
@@ -14,7 +14,7 @@ class MeetingBase(BaseModel):
     date: str
     employeeNumber: int
     time: int
-    totalCost: int
+    totalCost: float 
     powerpoint: bool
     powerpointSlides: int
     comment: str
